feat: add settings module with environment modes
