+++ apps/backend/app/core/app_settings/embedding.py
