+++ apps/backend/app/core/app_settings/navigation.py
